docs: update `with-typescript` example using the new `createClient()` type parameter
