+++ examples/with-typescript/index.ts
@@ -1,87 +1,38 @@
 import * as prismic from "@prismicio/client";
-import * as prismicT from "@prismicio/types";
 import fetch from "node-fetch";
 
-/**
- * A Page document from Prismic.
- */
-type PrismicDocumentPage = prismicT.PrismicDocument<
-	{
-		title: prismicT.KeyTextField;
-		footnotes: prismicT.RichTextField;
-	},
-	"page",
-	"en-us" | "fr-fr"
->;
-
-/**
- * A Blog Post document from Prismic. This example contains Slices.
- */
-type PrismicSliceBlogPostBodyText = prismicT.Slice<
-	"text",
-	never,
-	{
-		text: prismicT.RichTextField;
-	}
->;
-type PrismicSliceBlogPostBodyImageGallery = prismicT.Slice<
-	"image_gallery",
-	never,
-	{
-		image: prismicT.ImageField;
-		caption: prismicT.KeyTextField;
-	}
->;
-type PrismicSliceBlogPostBodyQuote = prismicT.Slice<
-	"quote",
-	{
-		quote: prismicT.RichTextField;
-		quotee: prismicT.KeyTextField;
-	}
->;
-type PrismicDocumentBlogPost = prismicT.PrismicDocument<
-	{
-		title: prismicT.KeyTextField;
-		description: prismicT.RichTextField;
-		body: prismicT.SliceZone<
-			| PrismicSliceBlogPostBodyText
-			| PrismicSliceBlogPostBodyImageGallery
-			| PrismicSliceBlogPostBodyQuote
-		>;
-	},
-	"page",
-	"en-us" | "fr-fr"
->;
+// See `./types.ts` for the document types.
+import { AllDocumentTypes } from "./types";
 
 const main = async () => {
-	const client = prismic.createClient("qwerty", { fetch });
+	// You can pass a union of document types to `createClient()`.
+	// All query methods will now be typed appropriately. Methods that
+	// accept a document type, for example, will return the correct type
+	// for that document type.
+	const client = prismic.createClient<AllDocumentTypes>("qwerty", { fetch });
 
-	// We can pass a type to client methods that return documents.
-	// Here, we tell `client.getByUID` that the document type will be a `PrismicDocumentPage`.
-	const homepage = await client.getByUID<PrismicDocumentPage>("page", "home");
-	// => The Page document with a UID of `home`
+	const homepage = await client.getByUID("page", "home");
+	//    ^ Typed as PageDocument
 
 	const type = homepage.type;
-	// => Typed as `"page"`
+	//                    ^ Typed as "page"
 	console.info("title: ", type);
 
 	const lang = homepage.lang;
-	// => Typed as `"en-us" | "fr-fr"`
+	//                    ^ Typed as `"en-us" | "fr-fr"`
 	console.info("lang: ", lang);
 
 	const title = homepage.data.title;
-	// => Typed as `string | null`
+	//                          ^ Typed as `string | null`
 	console.info("title: ", title);
 
 	// @ts-expect-error - non_existant_field does not exist in `data`
 	const nonExistantField = homepage.data.non_existant_field;
-	// TypeScript Error: Property 'non_existant_field' does not exist on type '{ title: KeyTextField; }'
+	// TypeScript Error: Property `non_existant_field` does not exist on type `PageDocument['data']`
 	console.info("nonExistantField: ", nonExistantField);
 
-	const blogPosts = await client.getAllByType<PrismicDocumentBlogPost>(
-		"blog_post",
-	);
-	// => All Blog Post documents.
+	const blogPosts = await client.getAllByType("blog_post");
+	//    ^ Typed as BlogPostDocument[]
 	console.info("blogPosts: ", blogPosts);
 };
